+++ src/asp/Camera/LinescanSpotModel.cc
@@ -252,14 +252,16 @@ boost::shared_ptr<SPOTCameraModel> load_spot5_camera_model_from_xml(std::string 
   // This is where we could set the Earth radius and mean surface elevation if we have that info.
 
   // Feed everything into a new camera model.
+  bool enable_correct_velocity_aberration = false;
+  bool enable_correct_atmospheric_refraction = false;
   return boost::shared_ptr<SPOTCameraModel>
     (new SPOTCameraModel(position_func, velocity_func, 
                          pose_func, time_func, 
                          xml_reader.look_angles,
                          xml_reader.image_size,
                          min_time, max_time,
-                         stereo_settings().enable_correct_velocity_aberration,
-                         stereo_settings().enable_correct_atmospheric_refraction));
+                         enable_correct_velocity_aberration,
+                         enable_correct_atmospheric_refraction));
   
 } // End function load_spot_camera_model()