+++ src/asp/Tools/bundle_adjust.h
@@ -15,7 +15,6 @@
 //  limitations under the License.
 // __END_LICENSE__
 
-
 #ifndef __ASP_TOOLS_BUNDLEADJUST_H__
 #define __ASP_TOOLS_BUNDLEADJUST_H__
 
@@ -77,9 +76,7 @@ struct Options: public asp::BaBaseOptions {
     reference_terrain_weight, auto_overlap_buffer, pct_for_overlap;
   bool skip_rough_homography, enable_rough_homography, disable_tri_filtering,
     enable_tri_filtering, no_datum, individually_normalize, use_llh_error,
-    force_reuse_match_files, save_cnet_as_csv,
-    enable_correct_velocity_aberration, enable_correct_atmospheric_refraction, 
-    dg_use_csm, aster_use_csm;
+    force_reuse_match_files, save_cnet_as_csv, aster_use_csm;
   vw::Vector2 elevation_limit;   // Expected range of elevation to limit results to.
   vw::BBox2 lon_lat_limit;       // Limit the triangulated interest points to this lonlat range
   vw::BBox2 proj_win; // Limit input triangulated points to this projwin
@@ -112,11 +109,6 @@ struct Options: public asp::BaBaseOptions {
     asp::stereo_settings().num_scales                 = num_scales;
     asp::stereo_settings().nodata_value               = nodata_value;
     
-    asp::stereo_settings().enable_correct_atmospheric_refraction
-      = enable_correct_atmospheric_refraction;
-    asp::stereo_settings().enable_correct_velocity_aberration
-      = enable_correct_velocity_aberration;
-    asp::stereo_settings().dg_use_csm = dg_use_csm;
     asp::stereo_settings().aster_use_csm = aster_use_csm;
     asp::stereo_settings().ip_per_tile = ip_per_tile;
     asp::stereo_settings().ip_per_image = ip_per_image;